Dryrun is now a boolean option to the call() method
Default value of dryRun is set to false.
Use setDryRun(true).call() to use

+++ org.eclipse.jgit/src/org/eclipse/jgit/api/CleanCommand.java
@@ -63,6 +63,8 @@ import org.eclipse.jgit.util.FileUtils;
 public class CleanCommand extends GitCommand<Set<String>> {
 
 	private Set<String> paths = Collections.emptySet();
+	
+	private boolean dryRun = false;
 
 	/**
 	 * @param repo
@@ -86,7 +88,8 @@ public class CleanCommand extends GitCommand<Set<String>> {
 			Status status = command.call();
 			for (String file : status.getUntracked()) {
 				if (paths.isEmpty() || paths.contains(file)) {
-					FileUtils.delete(new File(repo.getWorkTree(), file));
+					if (!dryRun)
+						FileUtils.delete(new File(repo.getWorkTree(), file));
 					files.add(file);
 				}
 			}
@@ -108,4 +111,16 @@ public class CleanCommand extends GitCommand<Set<String>> {
 		return this;
 	}
 
+	/**
+	 * If dryRun is set, the resources in question will not actually be deleted,
+	 * but a list of their names would be returned instead.
+	 *
+	 * @param dryRun
+	 *            the paths to set
+	 * @return {@code this}
+	 */
+	public CleanCommand setDryRun(boolean dryRun) {
+		this.dryRun = dryRun;
+		return this;
+	}
 }
